external modules: fixes in sql manipulations modules
