+++ external/etl/file/directory.py
@@ -3,7 +3,8 @@ from pathlib import Path
 from external.native.connections import get_engine
 from external.etl.file.excel import extract, is_valid
 from external.etl.sql import load_clean
-from external.utils.sql.common import dummy_db_action, truncate_table, apply_on_db
+from external.utils.sql.common import dummy_db_action, apply_on_db
+from external.utils.sql.delete import truncate_table
 from external.utils.sql.grant import grant_preset_priveleges
 from external.utils.sql.introspection import get_loaded_src_ids
 from external.utils.var import color
